+++ src/main/java/com/main/server/model/OrganizationPermissionType.java
@@ -1,4 +1,4 @@
-package com.main.server.model.organization;
+package com.main.server.model;
 
 public enum OrganizationPermissionType {
     DELETE("delete"),
