+++ src/main/java/com/main/server/service/impl/ProductServiceImpl.java
@@ -3,8 +3,8 @@ package com.main.server.service.impl;
 import com.main.server.exception.ResourceAlreadyExistException;
 import com.main.server.exception.ResourceNotFoundException;
 import com.main.server.model.User;
-import com.main.server.model.organization.Organization;
-import com.main.server.model.product.Product;
+import com.main.server.model.Organization;
+import com.main.server.model.Product;
 import com.main.server.repository.ProductRepository;
 import com.main.server.service.ProductService;
 import com.main.server.service.UserService;
